--- src/customChains.ts
@@ -1,15 +1,0 @@
-import { Chain, ChainContract, Client } from 'viem';
-
-export type CustomParentChain = Chain & {
-  contracts: {
-    rollupCreator: ChainContract;
-    tokenBridgeCreator: ChainContract;
-  };
-};
-
-export const customChains: Chain[] = [];
-
-export function registerCustomParentChain(chain: CustomParentChain) {
-  // todo: don't duplicate
-  customChains.push(chain);
-}
