+++ src/utils/getParentChainFromId.ts
@@ -1,14 +1,13 @@
 import { Chain, extractChain } from 'viem';
 
 import { validateParentChain } from '../types/ParentChain';
-import { chains } from '../chains';
-import { customChains } from '../customChains';
+import { chains, getCustomParentChains } from '../chains';
 
 export function getParentChainFromId(chainId: number): Chain {
   const { chainId: parentChainId } = validateParentChain(chainId);
 
   return extractChain({
-    chains: [...chains, ...customChains],
+    chains: [...chains, ...getCustomParentChains()],
     id: parentChainId,
   });
 }
